lines 154 and 170
The total number of combinations for two flips (lines 145-152) and for three flips (lines 157-168) differ from the total number of sampled coin flips (i.e., 10000). To get the proportions of combinations, we should divide combinations in the sequence by the total number of sequence (i.e., sum(table(sequences)). By dividing the numbers of combinations by the total number of combinations, proportions add up to 1.00