Line 164  z-score transformation formula
Hello Professor Crump, In your written and video lecture, you explained that "the centering transformation causes all values below the mean to receive a negative z-score, while all values above the mean to have a positive z-score."  According to the formula in our written lecture, the opposite is true (i.e., scores below the mean receive a positive number, and scores above the mean receive a negative number).   I adjusted the numerator of the formula, which I think is a typo.